added PostDAO, fixed ojdbc7 dependency, added access to database credentials
added a Post DAO (spring boot just needs an empty interface)
The dependency was in the wrong place, moved it to the right one
edited application.properties so that it will fetch the database credentials from environment variables. This will not build locally because we do not have access to the database on our local machines or the correct environment variables set, if anyone wants to test locally they need to make their own database and set the environment variables, and build from the command line because it does not seem to have access to environment variables when run from eclipse.